\name{FizzBuzz}
\alias{FizzBuzz}
\title{FizzBuzz}
\usage{
FizzBuzz(n)
}
\description{
Loops from 1 to n and returns a vector of elements based on the following rules
divisible by 3 → Fizz
divisible by 5 → Buzz
divisible by both → Fizz Buzz
divisible by neither → number
}
\examples{
FizzBuzz(50)
}
\param{n}{The upper limit of the loop.}
