% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{A time series of daily weather data in Flagstaff, AZ}
\format{
a data frame with 25449 observations
\describe{
  \item{DATE}{The date of observation in the format YYYY-MM-DD.}
  \item{SNOW}{The number of inches of snowfall for given day}
  \item{TMAX}{The maximum temperature recorded in Fahrenheit}
  \item{TMIN}{The minimum temperature recorded in Fahrenheit}
}
}
\source{
\url{www.ncdc.noaa.gov}
}
\usage{
Flagstaff_Weather
}
\description{
A time series of daily weather data in Flagstaff, AZ
}
\keyword{datasets}
